+++ changelog.md
@@ -27,6 +27,7 @@ Codion Change Log
 - EntityTablePanel.control() no longer Optional based, containsControl() added.
 - EntityEditPanel, EntityTablePanel.setupControls() added for setting up custom controls, called after standard controls have been initialized.
 - EntityDialogs.EditDialogBuilder.Updater added for customizing how the actual update is performed when multiple entities are edited, related refactoring.
+- EntityTablePanel.editDialogBuilder() added.
 
 ## 0.17.21
 ### is.codion.swing.common.model